Fix horizontal scrolling issues.
The problem was the attempt to move the viewport. One should move instead the widget.
But there is no widget here, so we handle the horizontal scrolling like the verical.

Shift the painter and offset the screen coordinate of a click.

Reduce the maximum size of the horizontal scrollbar, so it is just enough to fit th width (like it is done for the vertical one).