Make QHexView an actual library.
As it does have sources, I don't think this is an interface. If not separate .so/.dll file is desired, this library can be linked statically.
But the way it was, it would have been recompiled in every other library that used the QHexView target. I currently don't see the sense in that.