Fixed last line being clipped on a system with high DPI monitor (issue 87)
Link to issue: https://github.com/Dax89/QHexView/issues/87

A few extra things to note:

1) Changed floating point arithmetic to integer arithmetic in QHexView::lines and QHexView::visibleLines.

For the former I am certain this is the correct choice because all arguments involved are integers. As for the latter, I think the integer division should be applied because the return value is integer. I beleive std::div should be reliable for this: the quotent tells us the number of visible lines and remainder marks a partially visible line.

2) I took the liberty of writing a few comments describing a few thoughts regarding the functions I encountered during making this fix, I hope you don't mind.